exagram/requirements.txt
Django>=5.1
djangorestframework
djangorestframework-simplejwt
django-debug-toolbar
django-cors-headers
django-extensions